+++ src/daf_gui/tod_extension.py
@@ -1,0 +1,3 @@
+"""
+Module that adds extensions to the existing tk-object-define library.
+"""
